requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0
openai>=1.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
